add LinksMapper and related tests; refactor UserMapper and service implementations
